Internal cleanup of TreehouseContent
Rename TreehouseContent to TreehouseAppContent. The name TreehouseContent
is already taken by Compose.

Rename RealBinding to ViewContentCodeBinding.

Move ViewContentCodeBinding into the same file as TreehouseAppContent.

Remove the unused type parameter on Content.
